reqres_tests/pyproject.toml
[project]
name = "reqres-tests"
version = "0.1.0"
description = "API testing project for ReqRes"
authors = [
    {name = "Tatiana ",email = "90062565+tatafitz@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "email-validator (>=2.1.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi-pagination (>=0.12.17,<0.13.0)"
]

[tool.poetry]
packages = [{include = "reqres_tests", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
